backend/requirements.txt
fastapi
uvicorn
pydantic
transformers
python-dotenv
torch
google-generativeai